Refactor Logic and Markdown view controllers
