+++ studio/Cartfile.resolved
@@ -1,7 +1,7 @@
 github "Lona/ColorPicker" "08ff046d9bcee4ad85581748fb02cce7497dec4a"
 github "Lona/ControlledComponents" "262f562cec208a96a6627857400205363429d6e0"
 github "Lona/FileTree" "cdfa6f8bb4be10e497ed0f76c5173c0495275474"
-github "Lona/Logic" "3387f3f8ac57ed70e0db283f99bcff7868fa2582"
+github "Lona/Logic" "3835ac1f13cb4b352737408fc96340b0ac56c232"
 github "davecom/SwiftGraph" "3.0.0"
 github "dylan/colors" "0.9.10"
 github "njdehoog/Witness" "0.1.3"
