+++ studio/LonaStudio/AppDelegate.swift
@@ -114,7 +114,7 @@ class AppDelegate: NSObject, NSApplicationDelegate {
         if let viewControllers = welcomeWindow.contentViewController?.presentedViewControllers {
             viewControllers.forEach { $0.dismiss(sender) }
         }
-        
+
         welcomeWindow.close()
     }
